python-ai-service/requirements.txt
# FastAPI and ASGI server
fastapi
uvicorn

# Environment variable loading
python-dotenv

# LangChain core and provider integrations
langchain
langchain-google-genai
langchain-openai